removed old mmr proof
