+++ pallets/bridge-common/Cargo.toml
@@ -37,11 +37,11 @@ sp-io = { git = "https://github.com/paritytech/substrate.git", branch = "polkado
 sp-core = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.38", default-features = false }
 sp-beefy = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.38", default-features = false }
 bridge-types = { path = "../types", default-features = false }
+sp-runtime = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.38", default-features = false }
 
 [dev-dependencies]
 # Substrate
 sp-core = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.38", default-features = false }
-sp-runtime = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.38", default-features = false }
 hex-literal = "0.3.1"
 serde_json = "1.0.73"
